clb-384-bug updateSslTermination logs now accurately report when the non secure virtual server is enabled/disabled.
clb-385 and clb-386 bug fixes.  Usage is collected before delete ssl termination and after.  When an event happens and no records are returned by SNMP, those records are now inserted with -1 anyway.

A record will be created for every host on a cluster for every event that comes in regardless of whether or not it can be retrieved by snmp or not.

Added log statements for what the exact data is being inserted into the database on events.  This helps with debugging.