Update patch set 3
Patch Set 3:

didnt have amq running and it got stuck in pending update. But when trying to make subsequent request(once amq is started) I get.

CertificateMappingServiceImpl.setLbToPendingUpdate()] Load Balancer '3108' has a status of 'null' and is considered immutable.

Which probably means the status check in this piece of code is broken. (not looking/retrieving proper status of the lb)

The database is showing the correct status.

Patch-set: 3