Update patch set 3
Patch Set 3: Code-Review-1

Another issues(which may or may not happen in production and could cause issues)

I had tested before the zxtm capability was added. Thus, my cert mappings lived in the DB. When attempting to remove the mappings zeus fails with:

The SSL site with IP/hostname '*.jorge.com' could not be found.

This will cause all kinds of failures with sync/out of sync load balancers.

maybe still allow the database call to complete succesfully if there is not a matching cert in zeus. (This would be expected behaviour in this cases IMO)

Patch-set: 3
Label: Code-Review=-1