Update patch set 4
Patch Set 4: Code-Review+1 Verified+1

I'd like to get this merged so I can run it up to Staging for testing ASAP. I've run a full regression suite on it in Dev and it comes back clean. Running the suite again now to make sure, but I think merging this should be a no-brainer (unless we want to merge it as 1.25 or something).

Patch-set: 4
Label: Code-Review=+1
Label: Verified=+1