Start nodes ONLINE regardless of condition
If a node is DRAINING or DISABLED, that doesn't necessarily affect its
availability. Only a health monitor should set nodes to OFFLINE, because
only a health monitor can set a node back to ONLINE.

Updating unit tests for this, and an unrelated failure due to DNS
resolution of our "bad ipv4 address", which happens to be a valid domain.

Change-Id: Ia1ca2fc3834822c6a80ed2298129c675b931b29e
Closes-Bug: CLB-8 (was CLB-750)