Update patch set 1
Patch Set 1:

VSSTOP Issue: Traffic IP Groups are starting, then quickly stopping on the new clusters, meaning they stay down.

I am fairly certain this is a config replication issue, triggered by making two very quick successive calls to the RESTAPI to create the TIGs first in OFFLINE, then ONLINE them -- the replication is causing them to be interpreted backwards, meaning it is first created in ONLINE, then updated to OFFLINE. :(

Due to the speed at which we've demonstrated in our runs, it is not actually necessary to "pre-create" the objects in a disabled state (it was supposed to save time, but really doesn't and causes this problem).

Patch-set: 1