Update patch set 1
Patch Set 1: Code-Review-1

I want the full stack trace for all the exceptions. to get the full message
import org.openstack.atlas.util.debug.Debug

try{
}catch(Exception ex){
    LOG.error(String.format("Exception: %s", Debug.getExtendedStackTrace(ex));}
}

e.message usually only gives the message body of the exception. Also I want the full stack trace of all exceptions in the catch block and not just instances of AxisFault. The actual message of the stack trace can be generic if you want. Its the stack trace itself that will tell us the path the code took when it hit the error. There are also places in the code where exception handling seems to be done which I'm sure is to hide default cases.

Patch-set: 1
Label: Code-Review=-1