Update patch set 1
Patch Set 1:

I only bring this up here, and am a bit bothered because it seems like we're not on the same page. It's getting quite difficult moving forward and we need to ensure we're on the same path to get there. If we dont want to use gerrit then so be it. I'd really rather not pass around patch files though. Can we please come to a consensus and start making things move forward?

Patch-set: 1