Update patch set 1
Patch Set 1: Code-Review+2

This should still work in our 1.8 efforts? Will we just leave it alone or do the same thing in the new stuff? 

We shouldn't need to going forward... hopefully, but if we do change those schemas we need to make sure to update this in 1.6. Just as a clarification.

Patch-set: 1
Reviewer: Phil Toohill <1001832@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2