Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)
> 
> We should verify our new functions are called when they should be

This test was failing with NPE as there is no protocol on the load balancer used in the test. The protocol is a required field on a load balancer. The update load balancer operation expects the required fields to be initialized on the load balancer.

Will look into the test again.

Patch-set: 1
Reviewer: Gerrit User 1010948 <1010948@d5d70762-12d0-45a1-890d-524b12d3f735>